Renaming some classes and changes on footer class
Renamed the file from "Press" to "press" and made minor changes to the footer. The 'social media' class has been renamed to 'follow-on-social-media,' and the 'social links' class has been renamed to 'social-media.' The 'footer' class has been renamed to 'footer-container,' and the old 'footer' class called 'site-footer' has been removed.

All <div> elements previously used in the footer sections have been replaced with <nav> elements. The class name 'footer-logo-list-item' and the <a> element now include the link to the image displayed for the link. Previously, the footer logos were individually defined and styled in the CSS file. Now, the images are included directly in the HTML file, which makes the code cleaner and more understandable. All logos can be managed and adjusted using the 'footer-logo-item' class. This change also simplifies the use of the 'alt' attribute for the logos, enhancing accessibility.